feat: integrate RandomStringGenerator into AegisCrypto and remove standalone class
